+++ packages/account/src/Components/file-uploader-container/file-uploader.tsx
@@ -1,4 +1,4 @@
-//TODO this file has types error, needed to be refactored. typing here skiping checks for build passing
+//TODO all file upload process needed to be checked and refactored with TS. skipping checks for passing CFD build
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-nocheck
 import React from 'react';
