+++ packages/account/src/Sections/Verification/ProofOfAddress/proof-of-address.tsx
@@ -22,6 +22,4 @@ const ProofOfAddress = observer(() => {
     );
 });
 
-ProofOfAddress.displayName = 'ProofOfAddress';
-
 export default ProofOfAddress;
