+++ packages/account/src/Sections/Verification/ProofOfIdentity/proof-of-identity-container-for-mt5.jsx
@@ -6,7 +6,7 @@ import ErrorMessage from 'Components/error-component';
 import IconWithMessage from 'Components/icon-with-message';
 import POISubmissionForMT5 from './proof-of-identity-submission-for-mt5.jsx';
 import { service_code } from './proof-of-identity-utils';
-import { populateVerificationStatus } from '../Helpers/verification';
+import { populateVerificationStatus } from '../Helpers/verification.js';
 
 const ProofOfIdentityContainerForMt5 = ({
     account_settings,
